requirements.txt
discord.py
flask
yt-dlp
pynacl